+++ spec/olels/bib_editor_spec.rb
@@ -35,6 +35,8 @@ describe 'The Bib Editor' do
   it 'should have bib editor elements' do
     elements = @bib_editor.elements
     elements.include?(:message).should be_true
+    elements.include?(:messages).should be_true
+    elements.include?(:message_header).should be_true
     elements.include?(:leader_field).should be_true
     elements.include?(:bib_record_status_selector).should be_true
     elements.include?(:set_button).should be_true
@@ -56,6 +58,11 @@ describe 'The Bib Editor' do
     elements.include?(:control_007_set_button).should be_true
     elements.include?(:control_007_reset_button).should be_true
     elements.include?(:control_008_field).should be_true
+  end
+
+  it 'should have bib editor functions' do
+    functions = @bib_editor.functions
+    functions.should include(:message_count)
   end
 
   it 'should start with one data line added' do