+++ CHANGELOG.md
@@ -1,5 +1,10 @@
 ### v2.5.1 - 2013/08/??
 
+* Small Fixes
+    * Update spec/common/page_spec.rb
+    * Update specs to ensure @ole.quit
+      is run after each full spec
+
 ### v2.5.0 - 2013/08/05
 
 * Add Patron Record Screen